daemon: move `GetServiceList` to loadbalancer.go
Signed-off by: Ian Vernon <ian@cilium.io>
